migrate TrajectoryChart
